+++ liana/method/sp/_Misty.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-import logging
+from liana._logging import _logg
 from tqdm import tqdm
 
 from scipy.sparse import isspmatrix_csr, csr_matrix
@@ -10,7 +10,6 @@ from sklearn.linear_model import LinearRegression, RidgeCV
 from sklearn.model_selection import KFold, cross_val_predict
 import statsmodels.api as sm
 
-from anndata import AnnData
 from mudata import MuData
 
 class MistyData(MuData):
@@ -32,7 +31,6 @@ class MistyData(MuData):
         **kwargs : `dict`
             Keyword arguments passed to the MuData Super class
         """
-        
         if isinstance(data, MuData):
             data = data.mod
         
@@ -48,13 +46,12 @@ class MistyData(MuData):
         
         for view in self.view_names:
             if not isspmatrix_csr(self.mod[view].X):
-                 logging.warning(f"view {view} is not a csr_matrix. Converting to csr_matrix")
-                 self.mod[view].X = csr_matrix(self.mod[view].X)
+                _logg(f"view {view} is not a csr_matrix. Converting to csr_matrix", verbose=True, level='warn')
+                self.mod[view].X = csr_matrix(self.mod[view].X)
             if view=="intra":
                 continue
             if f"{self.spatial_key}_connectivities" not in self.mod[view].obsp.keys():
                 raise ValueError(f"view {view} does not contain `{self.spatial_key}_connectivities` key in .obsp")
-
     
     def _get_conn(self, view_name):
         return self.mod[view_name].obsp[f"{self.spatial_key}_connectivities"]
@@ -65,8 +62,7 @@ class MistyData(MuData):
                  predict_self = False,
                  k_cv = 10,
                  alphas = [0.1, 1, 10],
-                 intra_groupby = None,
-                 extra_groupby = None,
+                 maskby = None,
                  n_jobs = -1,
                  seed = 1337,
                  inplace=True,
@@ -74,7 +70,7 @@ class MistyData(MuData):
                  **kwargs
                  ):
         """
-        A Multi-view Learning for dissecting Spatial Transcriptomics data (MISTY) model.
+        A Multi-view Learning for dissecting Spatial Transcriptomics data (MISTy) model.
         
         Parameters
         ----------
@@ -83,20 +79,20 @@ class MistyData(MuData):
         model : `str`, optional (default: 'rf')
             Model used to model the single views. Default is 'rf'.
             Can be either 'rf' (random forest) or 'linear' (linear regression).
+        predict_self : `bool`, optional (default: False)
+            Whether to predict self-interactions. These are determined purely by the feature names.
         bypass_intra : `bool`, optional (default: False)
-            Whether to bypass modeling the intraview features importances via LOFO
-        intra_groupby : `str`, optional (default: None)
-            Column in the .obs attribute used to group cells in the intra-view
-            If None, all cells are considered as one group. (EXPERIMENTAL)
-        extra_groupby : `str`, optional (default: None)
-            Column in the .obs attribute used to group cells in the extra-view(s)
-            If None, all cells are considered as one group. (EXPERIMENTAL)
+            Whether to bypass modeling the intraview via LOFO.
+            In other words, whether to bypass modelling each target by leave-one-feature-out within the same spots.
+        maskby : `str`, optional (default: None)
+            Column in the .obs attribute used to group or mask observations in the intra-view
+            If None, all cells are considered as one group.
+        k_cv : `int`, optional (default: 10)
+            Number of folds for cross-validation used in the multi-view model, 
+            and single-view models if model is 'linear'.
         alphas : `list`, optional (default: [0.1, 1, 10])
             List of alpha values used to choose from, that control the strength of the ridge regression,
-            used for the multi-view part of the model
-        k_cv : `int`, optional (default: 10)
-            Number of folds for cross-validation used in the multi-view model, and single-view models if
-            model is 'linear'.
+            used for the multi-view part of the model. Only used if there are more than 2 views being modeled (including intra).
         n_jobs : `int`, optional (default: -1)
             Number of cores used to construct random forest models
         seed : `int`, optional (default: 1337)
@@ -104,6 +100,8 @@ class MistyData(MuData):
         inplace : `bool`, optional (default: True)
             Whether to write the results to the .uns attribute of the object or return 
             two DataFrames, one for target metrics and one for importances.
+        verbose : `bool`, optional (default: False)
+            Whether to show a progress bar.
         **kwargs : `dict`
             Keyword arguments passed to the Regressors. Note that n_jobs & random_state are already set.
         
@@ -113,121 +111,93 @@ class MistyData(MuData):
         Otherwise two DataFrames are returned, one for target metrics and one for importances.
 
         """
-        
-        # TODO: function that checks if the groupby is in the obs
-        # and does this for both extra & intra
-        intra_groups = np.unique(self.obs[intra_groupby]) if intra_groupby else [None]
-        extra_groups = np.unique(self.obs[extra_groupby]) if extra_groupby else [None]
-        
         view_str = list(self.view_names)
+        obs_masks = _create_obs_masks(self.mod['intra'], maskby)
         
         if bypass_intra:
             view_str.remove('intra')
         intra = self.mod['intra']
         
-        # init list to store the results for each intra group and env group as dataframe;
         targets_list, importances_list = [], []
         intra_features = intra.var_names.to_list()
-                
         progress_bar = tqdm(intra_features, disable=not verbose)
-        # loop over each target and build one RF model for each view
+        
         for target in (progress_bar):
-            if verbose:
-                progress_bar.set_description(f"Now learning: {target}")
-            
-            for intra_group in intra_groups:
-                intra_obs_msk = intra.obs[intra_groupby] == \
-                        intra_group if intra_group else np.ones(intra.shape[0], dtype=bool)
-                
-                # to array
-                y = intra[intra_obs_msk, target].X.toarray().reshape(-1)
-                # intra is always non-self, while other views can be self
-                predictors_nonself, insert_index = _get_nonself(target, intra_features)
-
-                # TODO: rename to target_importances
+            for intra_group in obs_masks.keys():
+                msk = obs_masks[intra_group]
                 importance_dict = {}
+                if verbose:
+                    d = f"Now learning: {target}" + \
+                        (f" masked by {intra_group}" if intra_group is not None else "")
+                    progress_bar.set_description(d)
+                    
+                predictors_nonself, insert_index = _get_nonself(target, intra_features)
+                y = intra[msk, target].X.toarray().reshape(-1)
+                X = intra[msk, predictors_nonself].X.toarray()
                 
-                # model the intraview
                 if not bypass_intra:
-                    predictions_intra, importance_dict["intra"] = _single_view_model(y,
-                                                                                     intra,
-                                                                                     intra_obs_msk,
-                                                                                     predictors_nonself,
-                                                                                     model=model,
-                                                                                     k_cv=k_cv,
-                                                                                     seed=seed,
-                                                                                     n_jobs=n_jobs,
-                                                                                     **kwargs
-                                                                                     )
+                    predictions_intra, importance_dict["intra"] = \
+                        _single_view_model(y,
+                                           X,
+                                           predictors_nonself,
+                                           model=model,
+                                           k_cv=k_cv,
+                                           seed=seed,
+                                           n_jobs=n_jobs,
+                                           **kwargs
+                                           )
                     if insert_index is not None and predict_self: 
                         # add self-interactions as nan
                         importance_dict["intra"][target] = np.nan
 
-                # loop over the group_views_by
-                for extra_group in extra_groups:
-                    # store the oob predictions for each view to construct predictor matrix for meta model
-                    predictions_list = []
+                # store the predictions for each view to construct predictor matrix for meta model
+                predictions_list = []
 
-                    if not bypass_intra:
-                        predictions_list.append(predictions_intra)
+                if not bypass_intra:
+                    predictions_list.append(predictions_intra)
 
-                    # model the juxta and paraview (if applicable)
-                    for view_name in [v for v in view_str if v != "intra"]:
-                        extra = self.mod[view_name]
-                        
-                        extra_features = extra.var_names.to_list()
-                        _predictors, _ =  _get_nonself(target, extra_features) if not predict_self else (extra_features, None)
-                        
-                        extra_obs_msk = self.obs[extra_groupby] == extra_group if extra_group is not None else None
-                        
-                        # NOTE: indexing here is expensive, but we do it to avoid memory issues
-                        connectivity = self._get_conn(view_name)
-                        view = _mask_connectivity(extra, connectivity, extra_obs_msk, _predictors)
-                        
-                        predictions_extra, importance_dict[view_name] = \
-                            _single_view_model(y,
-                                               view,
-                                               intra_obs_msk,
-                                               _predictors, 
-                                               model=model,
-                                               k_cv=k_cv,
-                                               seed=seed,
-                                               n_jobs=n_jobs,
-                                               **kwargs
-                                               )
-                        predictions_list.append(predictions_extra)
-
-                    # train the meta model with k-fold CV
-                    intra_r2, multi_r2, coefs = _multi_model(y,
-                                                             np.column_stack(predictions_list),
-                                                             intra_group,
-                                                             bypass_intra,
-                                                             view_str,
-                                                             k_cv,
-                                                             alphas, 
-                                                             seed
-                                                             )
+                # model the juxta and paraview (if applicable)
+                for view_name in [v for v in view_str if v != "intra"]:
+                    extra = self.mod[view_name]
                     
-                    # write the results to a dataframe
-                    targets_df = _format_targets(target,
-                                                 intra_group,
-                                                 extra_group,
-                                                 view_str,
-                                                 intra_r2,
-                                                 multi_r2,
-                                                 coefs
-                                                 )
-                    targets_list.append(targets_df)
+                    extra_features = extra.var_names.to_list()
+                    _predictors, _ =  _get_nonself(target, extra_features) if not predict_self else (extra_features, None)
                     
-                    importances_df = _format_importances(target=target, 
-                                                         intra_group=intra_group, 
-                                                         extra_group=extra_group,
-                                                         importance_dict=importance_dict
-                                                         )
-                    importances_list.append(importances_df)
+                    # NOTE: we multiply before masking
+                    weights = self._get_conn(view_name)
+                    X = weights @ extra[:, _predictors].X.toarray()
+                    X = X[msk, :]
+                    
+                    predictions_extra, importance_dict[view_name] = \
+                        _single_view_model(y,
+                                           X,
+                                           _predictors, 
+                                           model=model,
+                                           k_cv=k_cv,
+                                           seed=seed,
+                                           n_jobs=n_jobs,
+                                           **kwargs
+                                           )
+                    predictions_list.append(predictions_extra)
 
+                target_metrics = _multi_model(y,
+                                              np.column_stack(predictions_list),
+                                              intra_group,
+                                              bypass_intra,
+                                              view_str,
+                                              target,
+                                              k_cv,
+                                              alphas, 
+                                              seed
+                                              )
+                targets_list.append(target_metrics)
+                
+                importances_df = _format_importances(target=target,
+                                                     intra_group=intra_group, 
+                                                     importance_dict=importance_dict
+                                                     )
+                importances_list.append(importances_df)
 
-        # create result dataframes
         target_metrics, importances = _concat_dataframes(targets_list,
                                                          importances_list,
                                                          view_str)
@@ -242,11 +212,9 @@ class MistyData(MuData):
 def _create_dict(**kwargs):
     return {k: v for k, v in kwargs.items() if v is not None}
 
-def _format_targets(target, intra_group, extra_group, view_str, intra_r2, multi_r2, coefs):
-    # TODO: Remove dot from column names
+def _format_targets(target, intra_group, view_str, intra_r2, multi_r2, coefs):
     d = _create_dict(target=target,
                      intra_group=intra_group,
-                     extra_group=extra_group,
                      intra_R2=intra_r2,
                      multi_R2=multi_r2,
                      gain_R2=multi_r2 - intra_r2,
@@ -258,24 +226,23 @@ def _format_targets(target, intra_group, extra_group, view_str, intra_r2, multi_
     return target_df
 
 
-def _format_importances(target, intra_group, extra_group, importance_dict):
+def _format_importances(target, intra_group, importance_dict):
     
     importances_df = pd.DataFrame(importance_dict).reset_index().rename(columns={'index': 'predictor'})
-    importances_df[['target', 'intra_group', 'extra_group']] = target, intra_group, extra_group
+    importances_df[['target', 'intra_group']] = target, intra_group
         
     return importances_df
 
 
 def _concat_dataframes(targets_list, importances_list, view_str):
     target_metrics = pd.concat(targets_list, axis=0, ignore_index=True)
-    
-    target_metrics.loc[:, view_str] = target_metrics.loc[:, view_str].clip(lower=0)
-    target_metrics.loc[:, view_str] = target_metrics.loc[:, view_str].div(target_metrics.loc[:, view_str].sum(axis=1), axis=0)
-    
     importances = pd.concat(importances_list, axis=0, ignore_index=True)
     importances = pd.melt(importances,
-                          id_vars=["target", "predictor", "intra_group", "extra_group"], 
-                          value_vars=view_str, var_name="view", value_name="importances")
+                          id_vars=["target", "predictor", "intra_group"],
+                          value_vars=view_str, 
+                          var_name="view",
+                          value_name="importances"
+                          )
     
     # drop intra and extra group columns if they are all None
     importances = importances.dropna(axis=1, how='all')
@@ -284,29 +251,26 @@ def _concat_dataframes(targets_list, importances_list, view_str):
     return target_metrics, importances
 
 
-def _single_view_model(y, view, intra_obs_msk, predictors, model, k_cv, seed, n_jobs, **kwargs):
-    X = view[intra_obs_msk, predictors].X.toarray()
-    
+def _single_view_model(y, X, predictors, model, k_cv, seed, n_jobs, **kwargs):
     if model=='rf':
         model = RandomForestRegressor(oob_score=True,
                                       n_jobs=n_jobs,
                                       random_state=seed,
                                       **kwargs,
                                       )
-        # Model is a RandomForestRegressor
         model = model.fit(y=y, X=X)
         predictions = model.oob_prediction_
         importances = model.feature_importances_
         
     elif model=='linear':
         model = LinearRegression(n_jobs=1, **kwargs)
-        predictions = cross_val_predict(model, X, y,
+        predictions = cross_val_predict(model,
+                                        X, y,
                                         cv=KFold(n_splits=k_cv,
                                                  random_state = seed,
                                                  shuffle=True),
                                         n_jobs=n_jobs
                                         )
-        # NOTE: I use ols t-values for feature importances
         importances = sm.OLS(y, X).fit().tvalues
         
     else:
@@ -317,47 +281,60 @@ def _single_view_model(y, view, intra_obs_msk, predictors, model, k_cv, seed, n_
     return predictions, named_importances
 
 
-def _multi_model(y, predictions, intra_group, bypass_intra, view_str, k_cv, alphas, seed):
-    if predictions.shape[0] < k_cv:
-        logging.warn(f"Number of samples in {intra_group} is less than k_cv. "
-                     "{intra_group} values set to NaN")
-        return np.nan, np.nan, np.repeat(np.nan, len(view_str))
+def _multi_model(y, predictions, intra_group, bypass_intra, view_str, target, k_cv, alphas, seed):
+    n_views = len(view_str)
+    
+    if (predictions.shape[0] < k_cv) or (y.var() == 0.0):
+        if predictions.shape[0] < k_cv:
+            error_message = (f"Number of samples is less than k_cv, {target} metrics set to NaN")
+        else:
+            error_message = (f"Variance of '{target}' is 0.0, metrics set to NaN")
+        
+        _logg(error_message, verbose=True, level='warn')
+        return _format_targets(target,
+                            intra_group,
+                            view_str,
+                            np.nan,
+                            np.nan,
+                            np.repeat(np.nan, n_views)
+                            )
         
     kf = KFold(n_splits=k_cv, shuffle=True, random_state=seed)
     R2_vec_intra, R2_vec_multi = np.zeros(k_cv), np.zeros(k_cv)
-    coef_mtx = np.zeros((k_cv, len(view_str)))
+    coef_mtx = np.zeros((k_cv, n_views))
+    
+    model = RidgeCV(alphas=alphas) if n_views > 2 else LinearRegression()
     
     for cv_idx, (train_index, test_index) in enumerate(kf.split(predictions)):
-        ridge_multi_model = RidgeCV(alphas=alphas).fit(X=predictions[train_index], y=y[train_index])
-        R2_vec_multi[cv_idx] = ridge_multi_model.score(X=predictions[test_index], y=y[test_index])
-        coef_mtx[cv_idx, :] = ridge_multi_model.coef_
+        multi_model = model.fit(X=predictions[train_index], y=y[train_index])
+        R2_vec_multi[cv_idx] = multi_model.score(X=predictions[test_index], y=y[test_index])
+        coef_mtx[cv_idx, :] = multi_model.coef_
 
-        if not bypass_intra: 
-            # NOTE: first column of obp is always intra only prediction if bypass_intra is False
-            obp_train = predictions[train_index, 0].reshape(-1, 1)
-            obp_test = predictions[test_index, 0].reshape(-1, 1)
+        if not bypass_intra:
+            pred_train = predictions[train_index, 0].reshape(-1, 1)
+            pred_test = predictions[test_index, 0].reshape(-1, 1)
             
-            ridge_intra_model = RidgeCV(alphas=alphas).fit(X=obp_train, y=y[train_index])
-            R2_vec_intra[cv_idx] = ridge_intra_model.score(X=obp_test, y=y[test_index])
-            
-        # TODO: misty's ridgeCV p-values (ridge::pvals) are calculated as in 
-        # https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-12-372#Sec2
-        # these are needed for the scaling of RF importances
+            intra_model = model.fit(X=pred_train, y=y[train_index])
+            R2_vec_intra[cv_idx] = intra_model.score(X=pred_test, y=y[test_index])
 
-    intra_r2 = R2_vec_intra.mean() if not bypass_intra else 0
+    # format R2s
+    intra_r2 = R2_vec_intra.mean().clip(min=0) if not bypass_intra else 0
+    multi_r2 = R2_vec_multi.mean().clip(min=0)
     
-    return intra_r2, R2_vec_multi.mean(), coef_mtx.mean(axis=0)
-
-
-def _mask_connectivity(adata, connectivity, extra_obs_msk, predictors):
+    # format coefficients
+    coefs = coef_mtx.mean(axis=0).clip(min=0)
+    coefs = coefs / coefs.sum()
     
-    weights = connectivity.copy()
-    if extra_obs_msk is not None:
-        weights[:, ~extra_obs_msk] = 0
-    X = weights @ adata[:, predictors].X
-    view = AnnData(X=X, obs=adata.obs, var=pd.DataFrame(index=predictors), dtype='float32')
-    
-    return view
+    # format metrics to a dataframe
+    target_metrics = _format_targets(target,
+                                     intra_group,
+                                     view_str,
+                                     intra_r2,
+                                     multi_r2,
+                                     coefs
+                                     )
+        
+    return target_metrics
 
 def _get_nonself(target, predictors):
     if target in predictors:
@@ -368,3 +345,20 @@ def _get_nonself(target, predictors):
         predictors_subset = predictors
         insert_idx = None
     return predictors_subset, insert_idx
+
+def _create_obs_masks(intra, maskby):
+    obs_masks = {}
+    # if maskby is a column of only boleans take it as is    
+    if maskby is None:
+        obs_masks[None] = np.ones(intra.shape[0], dtype=bool)
+    elif intra.obs[maskby].dtype == bool:
+        obs_masks[None] = intra.obs[maskby]
+    # else if maskby is column of strings convert to categorical
+    elif intra.obs[maskby].dtype == 'category':
+        for intra_group in intra.obs[maskby].cat.categories:
+            obs_masks[intra_group] = intra.obs[maskby] == intra_group
+    else:
+        raise ValueError(f"maskby column {maskby} must be a column of booleans or categorical")
+        
+        
+    return obs_masks