+++ liana/multi/__init__.py
@@ -1,4 +1,4 @@
 from liana.multi.df_to_lr import df_to_lr
 from liana.multi.to_tensor_c2c import to_tensor_c2c
-from liana.multi.to_mudata import adata_to_views, lrs_to_views
+from liana.multi.to_mudata import adata_to_views, lrs_to_views, filter_view_markers
 from liana.multi._nmf import nmf, estimate_elbow